iASL: only point Node->Op to parse objects that represent elements in the namespace
This commit changes the namespace loading behavior to only assign
Node->Op = Op only if Op represents parse objects that create
elements in the namespace. In particular, this patch covers 2 cases:

case 1: Op is a scope operator.

case 2: Op is an external declaration of a named object and the named
        object definition existed prior to namespace loading.

In either case, the Op that is being loaded should not create a namespace node.
Therefore, Node->Op should not be re-assinged in these cases.

Signed-off-by: Erik Kaneda <erik.kaneda@intel.com>